\name{NIGp3}
\alias{NIGp3}
\title{
Normal inverse Gaussian process
}
\description{
Normal inverse Gaussian process. Not meant for use outside of other Bclim functions.
}
\usage{
NIGp3(begin, alpha, delta, totalCount, distance)
}
\arguments{
  \item{begin}{
Starting value
}
  \item{alpha}{
IG parameter
}
  \item{delta}{
IG parameter
}
  \item{totalCount}{
Total Count
}
  \item{distance}{
Distance
}
}
\details{
Not meant for us outside of other Bclim functions. See \code{\link{Bclim}} for information.
}
\value{
A set of values from the required NIG process
}
\author{
Thinh K Doan
}
\seealso{
\code{\link{Bclim}}
}
\examples{
## See other Bclim functions
}
\keyword{ model }
\keyword{ multivariate }
\keyword{ smooth }
